requirements.txt
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
pyyaml==6.0.1
rank-bm25==0.2.2
openai==1.54.0
google-genai
python-dotenv==1.0.0
